requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
alembic==1.16.5
    # via mlflow
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via starlette
blinker==1.9.0
    # via flask
cachetools==5.5.2
    # via
    #   google-auth
    #   mlflow-skinny
    #   mlflow-tracing
certifi==2025.8.3
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   flask
    #   mlflow-skinny
    #   uvicorn
cloudpickle==3.1.1
    # via mlflow-skinny
colorama==0.4.6
    # via click
contourpy==1.3.3
    # via matplotlib
cryptography==45.0.7
    # via mlflow
cycler==0.12.1
    # via matplotlib
databricks-sdk==0.65.0
    # via
    #   mlflow-skinny
    #   mlflow-tracing
docker==7.1.0
    # via mlflow
fastapi==0.116.1
    # via mlflow-skinny
flask==3.1.2
    # via mlflow
fonttools==4.59.2
    # via matplotlib
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via mlflow-skinny
google-auth==2.40.3
    # via databricks-sdk
graphene==3.4.3
    # via mlflow
graphql-core==3.2.6
    # via
    #   graphene
    #   graphql-relay
graphql-relay==3.2.0
    # via graphene
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
importlib-metadata==8.7.0
    # via
    #   mlflow-skinny
    #   opentelemetry-api
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
joblib==1.5.2
    # via scikit-learn
kiwisolver==1.4.9
    # via matplotlib
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   mako
    #   werkzeug
matplotlib==3.10.6
    # via mlflow
mlflow==3.3.2
    # via -r requirements.in
mlflow-skinny==3.3.2
    # via mlflow
mlflow-tracing==3.3.2
    # via mlflow
numpy==2.3.3
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   mlflow
    #   pandas
    #   scikit-learn
    #   scipy
opentelemetry-api==1.37.0
    # via
    #   mlflow-skinny
    #   mlflow-tracing
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-sdk==1.37.0
    # via
    #   mlflow-skinny
    #   mlflow-tracing
opentelemetry-semantic-conventions==0.58b0
    # via opentelemetry-sdk
packaging==25.0
    # via
    #   matplotlib
    #   mlflow-skinny
    #   mlflow-tracing
pandas==2.3.2
    # via
    #   -r requirements.in
    #   mlflow
pillow==11.3.0
    # via matplotlib
protobuf==6.32.1
    # via
    #   mlflow-skinny
    #   mlflow-tracing
pyarrow==21.0.0
    # via mlflow
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23
    # via cffi
pydantic==2.11.9
    # via
    #   fastapi
    #   mlflow-skinny
    #   mlflow-tracing
pydantic-core==2.33.2
    # via pydantic
pyparsing==3.2.4
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   graphene
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
pywin32==311
    # via docker
pyyaml==6.0.2
    # via mlflow-skinny
requests==2.32.5
    # via
    #   databricks-sdk
    #   docker
    #   mlflow-skinny
rsa==4.9.1
    # via google-auth
scikit-learn==1.7.2
    # via
    #   -r requirements.in
    #   mlflow
scipy==1.16.2
    # via
    #   mlflow
    #   scikit-learn
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via
    #   alembic
    #   mlflow
sqlparse==0.5.3
    # via mlflow-skinny
starlette==0.47.3
    # via fastapi
threadpoolctl==3.6.0
    # via scikit-learn
typing-extensions==4.15.0
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   graphene
    #   mlflow-skinny
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   docker
    #   requests
uvicorn==0.35.0
    # via mlflow-skinny
waitress==3.0.2
    # via mlflow
werkzeug==3.1.3
    # via flask
zipp==3.23.0
    # via importlib-metadata
